requirements.txt
Django==2.0.1
django-ckeditor==5.8.0
django-jet==1.0.8
django-js-asset==1.2.2
django-modeltranslation==0.14.2
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
